Cast away these warnings.

+++ src/Color.h
@@ -40,7 +40,7 @@ struct Color4ub {
 	Color4ub(Uint8 v_): r(v_), g(v_), b(v_), a(v_) {}
 	Color4ub(Uint8 r_, Uint8 g_, Uint8 b_): r(r_), g(g_), b(b_), a(255) {}
 	Color4ub(Uint8 r_, Uint8 g_, Uint8 b_, Uint8 a_): r(r_), g(g_), b(b_), a(a_) {}
-	Color4ub(const Color4f &c): r(c.r*255.f), g(c.g*255.f), b(c.b*255.f), a(c.a*255.f) {}
+	Color4ub(const Color4f &c): r(Uint8(c.r*255.f)), g(Uint8(c.g*255.f)), b(Uint8(c.b*255.f)), a(Uint8(c.a*255.f)) {}
 
 	operator unsigned char*() { return &r; }
 	operator const unsigned char*() const { return &r; }
